Edited file, fixed some errors in coding, will go over tomorrow
Chelsea,
I changed what I could on this file, I didn't want to get to far ahead on my changes w/o you involved.  There are some coding errors here that will not be hard to fix.  Some is just syntax, others is logical flow of the code.  We can work this out, and should be able to point you in the right direction.  

If you aren't tabbing your code over, it gets confusing.  I'm not sure if this is just lost in how you copied it over, or if your netbeans looks like this too, but we need to work on the format if that is the case.  It will help you read the code better, and things will start to click faster.